Refactor VarlenEntry out of storage_defs.h.
