+++ src/include/execution/sql/sql.h
@@ -5,6 +5,7 @@
 
 #include "common/macros.h"
 #include "execution/sql/runtime_types.h"
+#include "type/type_id.h"
 
 namespace noisepage::execution::sql {
